Refactor cheque write functionality and update PDF amount formatting
- Update CreateChequeWrite and EditChequeWrite to manage cheque status transitions.
- Change amount column type in migration to float for precision.
- Adjust PDF view to format cheque amount with two decimal places.

+++ database/migrations/2025_05_17_165704_create_cheque_writes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('cheque_id')->constrained()->cascadeOnDelete();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->string('payee');
-            $table->integer('amount');
+            $table->float('amount', 2);
             $table->date('date');
             $table->timestamps();
         });
